The README.md has been updated to include methods and results sections.
The results section needs to have expanded explainations, and the amount
of repeated code needs to be reduced by using functions that are called
for each feature.

All the data still needs to be moved over and the code needs the file
paths to be tidied up.

Lists of SRA IDs and genome IDs need to be added to the data.

Links/citations need to be added for all the code that is used.